Add files via upload
Problem Statement

The Indian Premier League, or IPL, is a T20 cricket league, which was founded in 2008 and is held every year. It sees participation from both national and international players, and eight teams representing eight Indian cities compete with each other in a double round-robin format in the league stages, which are followed by playoffs. Over the years, IPL has become one of the most-watched and most attended live sporting events all over the world.

 

You work as a data analyst at IFP, a nationally recognised news agency, which is based out of New Delhi, and provides news reports and feeds to magazines, newspapers and TV broadcasters all over the country. The Sports Editor of the agency has approached you to build a Tableau dashboard of IPL statistics over the years since its inception in order to create an infographic for a newsletter that their team is working on. For this newsletter, in some cases, they will use the visual representations as you have created in Tableau directly for their infographic, and in a few other cases, they will use important statistics after trying out the different filters and customisations that you have provided for interactivity. Therefore, you are expected to build an interactive dashboard in Tableau for this purpose. Let’s watch the upcoming video and listen to Shreyas as he states the broad objective of this assignment.

Play Video2079378
Play Video2079378
Play Video2079378
So, as Shreyas mentions in the video, in order to complete this assignment, you need to use the following two data set files:

matches.csv - It contains match-level information for each and every match held in IPL from 2008 to 2017.
deliveries.csv - It contains ball-by-ball information for each of the matches.

Combining the information in these two data sets, you will be creating an interactive dashboard that highlights some of the important statistics of IPL over the years. You will learn more about the datasets in one of the videos below.

 

You can download the data sets from the links provided below.
 

IPL Datasets
Download
 

Assignment Sub-tasks
In the upcoming video, Shreyas will explain the various sub-tasks that are associated with this entire dashboard building assignment.

So, as Shreyas mentions in the video, the final dashboard can be broken down into several categories and sub-categories. The ones which you need to focus on for this assignment are as follows:

 

1. Match Statistics

Toss outcome vs Match outcome (for each Ground/Venue)
Biggest wins (by runs and by wicket)
Highest totals (across all the seasons)
 

2. Player Statistics

Orange Cap contenders (The batsmen who have scored the maximum number of runs in a particular season)
Purple Cap contenders (The bowlers who have taken the maximum number of wickets in a particular season)
Batsmen who have hit the most number of fours and sixes (per season and overall)

3. Team Statistics

Season-wise team performance (wins vs losses)
Win %age ( home vs away)
 

Data Understanding
Let’s watch the upcoming video and listen to Shreyas as he describes the data set that you will be working on as part of the assignment.

So, as you saw in the video, the ‘deliveries.csv’ file contains ball-by-ball information for all the matches across all the seasons of IPL. Each row in this data set contains:

Match related information (batting team vs bowling team)
Player information (Bowler, Batsman, Non-striker)
Delivery information (Runs scored, Wickets, Extras, etc.)
 

Similarly, the ‘matches.csv’ file contains the following data:

Teams involved
Results (which team won, win type, player of the match, etc.)
Match specifics (umpires, ground, etc.)
 

Using these two datasets, you need to create visualisations pertaining to the different types of statistics mentioned in the assignment sub-tasks section. The final objectives and the expected results/submission are given below

 

Results Expected
A tableau workbook containing 3 dashboards pertaining to the 3 categories - Match statistics, Player statistics and Team statistics.
Each dashboard should contain the visualisations from the sub-category worksheets. For example, in the Match statistics dashboard, the visualisations should be from the Toss Outcome vs Match Outcome, Biggest Wins and Highest Totals worksheets. Similarly, you need to prepare the dashboards for Player statistics and Team statistics.
The dashboards and the corresponding visualisations should be well-detailed.
A PPT file containing the summary of the entire visual analysis for this assignment. This ppt should contain all the visualisations that have been used along with the necessary insights and details.